pyproject.toml
[project]
name = "vix-challenger"
version = "0.1.0"
description = "Compute a VIX-style 30-day implied volatility index from SPY options"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.0.0",
    "pyarrow>=15.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "kagglehub>=0.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "duckdb>=1.0.0",
    "numba>=0.59.0",
    "ipykernel>=6.29.0",
    "jupyterlab>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vix_challenger"]

